+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2018-11-23
 ### Added
+- `SQLiteDatabase#beginTransaction()` method
+- `SQLiteDatabase#beginTransactionNonExclusive()` method
+- `SQLiteDatabase#endTransaction()` method
 
 ## [0.1.0] - 2018-11-22
 ### Added